Focus on building a cli not a library
For a library it would be good to return a slice of IPs from ListCIDR
but if we do this then to print this slice out we would have to iterate
over it again, which slows down it's use as a tool. Unless there's a way
to do achieve both, say with a channel or a writer, we primarily want to
create a faster tool.